+++ src/extension.ts
@@ -100,7 +100,7 @@ function onDidChangeActiveTextEditor(editor: TextEditor | undefined) {
   const { languageId, uri } = editor.document;
 
   const workspace = clientWorkspaceForUri(uri, {
-    startIfMissing: languageId === 'rust' || languageId === 'toml',
+    initializeIfMissing: languageId === 'rust' || languageId === 'toml',
   });
   if (!workspace) {
     return;
@@ -108,11 +108,13 @@ function onDidChangeActiveTextEditor(editor: TextEditor | undefined) {
 
   activeWorkspace = workspace;
 
-  const updateProgress = (progress: { message: string } | null) => {
-    if (progress) {
+  const updateProgress = (progress: WorkspaceProgress) => {
+    if (progress.state === 'progress') {
       startSpinner(`[${workspace.folder.name}] ${progress.message}`);
     } else {
-      stopSpinner(`[${workspace.folder.name}]`);
+      const readySymbol =
+        progress.state === 'standby' ? '$(debug-stop)' : '$(debug-start)';
+      stopSpinner(`[${workspace.folder.name}] ${readySymbol}`);
     }
   };
 
@@ -139,12 +141,12 @@ function whenChangingWorkspaceFolders(e: WorkspaceFoldersChangeEvent) {
 const workspaces: Map<string, ClientWorkspace> = new Map();
 
 /**
- * Fetches a `ClientWorkspace` for a given URI. If missing and `startIfMissing`
+ * Fetches a `ClientWorkspace` for a given URI. If missing and `initializeIfMissing`
  * option was provided, it is additionally initialized beforehand, if applicable.
  */
 function clientWorkspaceForUri(
   uri: Uri,
-  options?: { startIfMissing: boolean },
+  options?: { initializeIfMissing: boolean },
 ): ClientWorkspace | undefined {
   const rootFolder = workspace.getWorkspaceFolder(uri);
   if (!rootFolder) {
@@ -157,14 +159,19 @@ function clientWorkspaceForUri(
   }
 
   const existing = workspaces.get(folder.uri.toString());
-  if (!existing && options && options.startIfMissing) {
+  if (!existing && options && options.initializeIfMissing) {
     const workspace = new ClientWorkspace(folder);
     workspaces.set(folder.uri.toString(), workspace);
-    workspace.start();
+    workspace.autoStart();
   }
 
   return workspaces.get(folder.uri.toString());
 }
+
+/** Denotes the state or progress the workspace is currently in. */
+type WorkspaceProgress =
+  | { state: 'progress'; message: string }
+  | { state: 'ready' | 'standby' };
 
 // We run one RLS and one corresponding language client per workspace folder
 // (VSCode workspace, not Cargo workspace). This class contains all the per-client
@@ -176,7 +183,7 @@ class ClientWorkspace {
   private readonly config: RLSConfiguration;
   private lc: LanguageClient | null = null;
   private disposables: Disposable[];
-  private _progress: Observable<{ message: string } | null>;
+  private _progress: Observable<WorkspaceProgress>;
   get progress() {
     return this._progress;
   }
@@ -185,11 +192,20 @@ class ClientWorkspace {
     this.config = RLSConfiguration.loadFromWorkspace(folder.uri.fsPath);
     this.folder = folder;
     this.disposables = [];
-    this._progress = new Observable<{ message: string } | null>(null);
+    this._progress = new Observable<WorkspaceProgress>({ state: 'standby' });
+  }
+
+  /**
+   * Attempts to start a server instance, if not configured otherwise via
+   * applicable `rust-client.autoStartRls` setting.
+   * @returns whether the server has started.
+   */
+  public async autoStart() {
+    return this.config.autoStartRls && this.start().then(() => true);
   }
 
   public async start() {
-    this._progress.value = { message: 'Starting' };
+    this._progress.value = { state: 'progress', message: 'Starting' };
 
     const serverOptions: ServerOptions = async () => {
       await this.autoUpdate();
@@ -265,6 +281,8 @@ class ClientWorkspace {
   public async stop() {
     if (this.lc) {
       await this.lc.stop();
+      this.lc = null;
+      this._progress.value = { state: 'standby' };
     }
 
     this.disposables.forEach(d => d.dispose());
@@ -308,9 +326,9 @@ class ClientWorkspace {
           } else if (progress.title) {
             status = `[${progress.title.toLowerCase()}]`;
           }
-          this._progress.value = { message: status };
+          this._progress.value = { state: 'progress', message: status };
         } else {
-          this._progress.value = null;
+          this._progress.value = { state: 'ready' };
         }
       },
     );
@@ -465,6 +483,14 @@ function registerCommands(): Disposable[] {
     commands.registerCommand(
       'rls.run',
       (cmd: Execution) => activeWorkspace && activeWorkspace.runRlsCommand(cmd),
+    ),
+    commands.registerCommand(
+      'rls.start',
+      () => activeWorkspace && activeWorkspace.start(),
+    ),
+    commands.registerCommand(
+      'rls.stop',
+      () => activeWorkspace && activeWorkspace.stop(),
     ),
   ];
 }